Update cachesimulator.py
Few issues here. First, convention is to use _ for an index that is unused. Second, the loop was adding "0" to the stringlength variable which is an integer. Next, the tag, set, and offsets were string splices of... nothing...? Should be splices of the binary string address. Not sure this is the proper way to do this. Need to calculate proper number of bits using formulas?