Update cachesimulator.py
Added m, t, b, and s from the cache formulas and used them to calculate which bits in the in the address will be tag, set, and offset bits. The program now changes these lengths dynamically based on the user-defined cache specifications. Next thing to implement is a better way to search through the cache. The current way is not correct since it searches for the tag as an index, not a "tag" is that makes sense.