Update cachesimulator.py
Soooo lots of changes here. Our program had all of our functions defined in main, which it should not have had. So I moved all the functions outside of main, which caused some scope errors. So I created global variables to hold all of the cache configuration variables. I also added dictionaries for the polciies so that we would be able to easily compute the name of the policy given the policy number. I also added in the eviction_line on a cache-read. This is always set to -1 currently, but I just wanted to see it in the output.