Update cachesimulator.py
Added in the 3rd extra credit replacment policy to the dictionary should we decide to do that. Also changed the cache_data to contain two single zeros at the beginning for the valid bit and the dirty bit. cache-view now prints these two bits at the beginning of each line. Next order of business is converting the tag bits to hex. They are currently in binary still.